+++ src/Providers/UserTenantProvider.php
@@ -3,12 +3,22 @@
 namespace MeghdadFadaee\NovaTenancy\Providers;
 
 use Illuminate\Contracts\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use LogicException;
 use MeghdadFadaee\NovaTenancy\Contracts\ProvidesNovaTenants;
 
 class UserTenantProvider extends EloquentTenantProvider
 {
+    public function find(Request $request, int|string $key): ?Model
+    {
+        if ($request->user() === null) {
+            return null;
+        }
+
+        return parent::find($request, $key);
+    }
+
     public function query(Request $request): Builder
     {
         $user = $request->user();